Implement select on server_sockfd
So far, the whole program is roughly completed. The debug message
should be removed.

Notice if you exit from play() and type new command, the command
may be wrong. The reason is that the output of the process may
pollute stdin, so the command that you type may have some un-
predictable prefix.

TODO: debug, improve readability and maintainability